Simle where args
