+++ apps/crawler/src/main.ts
@@ -8,7 +8,7 @@ async function bootstrap() {
       'log',
       'error',
       'warn',
-      // 'verbose',
+      'verbose',
       //
     ],
   });
