+++ apps/web-api/src/transaction/transaction.service.ts
@@ -7,7 +7,11 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { BaseService } from '../utils/base.service';
-import { IDataListResponse, IGQLQueryArgs } from '../utils/gql-query-args';
+import {
+  IDataListResponse,
+  IGQLQueryArgs,
+  TWhere,
+} from '../utils/gql-query-args';
 import { TransactionDTO } from './transaction.dto';
 
 @Injectable()
@@ -69,6 +73,8 @@ export class TransactionService extends BaseService<Event, TransactionDTO> {
       method: EventMethod.TRANSFER,
     });
 
+    this.applyWhereCondition(qb, queryArgs);
+
     this.applyOrderCondition(qb, queryArgs);
 
     this.applyLimitOffset(qb, queryArgs);
@@ -78,4 +84,14 @@ export class TransactionService extends BaseService<Event, TransactionDTO> {
 
     return { data, count };
   }
+
+  // applyWhereCondition(
+  //   qb: SelectQueryBuilder<T>,
+  //   args: IGQLQueryArgs<S>,
+  //   filterCb?: (
+  //     qb: SelectQueryBuilder<T>,
+  //     where: TWhere<S>,
+  //     method: OperatorMethods,
+  //   ) => void,
+  // ): void {}
 }
